Module Context.Proof

type merkle_leaf_kind =

| Hole

| Raw_context

type raw_context =

| Key of Tezos_base.TzPervasives.Bytes.t

| Dir of raw_context Tezos_base.TzPervasives.String.Map.t

| Cut

type merkle_hash_kind =

| Contents

| Node

type merkle_node =

| Hash of merkle_hash_kind * string

| Data of raw_context

| Continue of merkle_tree

and merkle_tree = merkle_node Tezos_base.TzPervasives.String.Map.t

val pp_raw_context : Stdlib.Format.formatter -> raw_context -> unit

type step = string

type value = bytes

type index = int

type hash = Tezos_base.TzPervasives.Context_hash.t

type !'a inode = {

length : int;

proofs : (index * 'a) list;

}

type !'a inode_extender = {

length : int;

segment : index list;

proof : 'a;

}

type tree =

| Value of value

| Blinded_value of hash

| Node of (step * tree) list

| Blinded_node of hash

| Inode of inode_tree inode

| Extender of inode_tree inode_extender

and inode_tree =

| Blinded_inode of hash

| Inode_values of (step * tree) list

| Inode_tree of inode_tree inode

| Inode_extender of inode_tree inode_extender

type kinded_hash = [

| `Node of hash

| `Value of hash

]

module Stream : sig ... end

Stream proofs represent an explicit traversal of a Merle tree proof. Every element (a node, a value, or a shallow pointer) met is first "compressed" by shallowing its children and then recorded in the proof.

type stream = Stream.t

type !'a t = {

version : int;

before : kinded_hash;

after : kinded_hash;

state : 'a;

}

module Internal_for_tests : sig ... end

val proof_hash_eq : 'a t -> 'b t -> bool