Module Dal_slot_repr.Index

An `Index.t` is a possible value for a slot index. We assume this value to be a positive 8-bit integer. Note that this is a hard constraint, which is independent of protocol constants. If a choice is ever made to increase the size of available slots in the protocol, we also need to change this module to accommodate for higher values.

type t

val encoding : t Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val pp : Tezos_protocol_environment_016_PtMumbai.Format.formatter -> t -> unit

val zero : t

val max_value : t

val of_int : int -> t option

of_int n constructs a`Slot_index.t`

val to_int : t -> int

val compare : t -> t -> int

val equal : t -> t -> bool