Module Tezos_raw_protocol_016_PtMumbai.Delegate_cycles

Per-cycle management of delegates.

val cycle_end : Raw_context.t -> Cycle_repr.t -> (Raw_context.t * Receipt_repr.balance_updates * Tezos_protocol_environment_016_PtMumbai.Signature.Public_key_hash.t list) Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Trigger the context maintenance at the end of cycle 'n', i.e.: unfreeze the endorsing rewards, potentially deactivate delegates. Return the corresponding balances updates and the list of deactivated delegates.

val init_first_cycles : Raw_context.t -> origin:Receipt_repr.update_origin -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

init_first_cycles ctxt ~origin computes and records the distribution of the total active stake among active delegates. This concerns the total active stake involved in the calculation of baking rights for all cycles in the range 0, preserved_cycles. It also freezes the deposits for all the active delegates.