Module Tezos_raw_protocol_016_PtMumbai.Frozen_deposits_storage

Simple abstraction from low-level storage to handle frozen deposits.

This module is responsible for maintaining the Storage.Contract.Frozen_deposits table.

val init : Raw_context.t -> Tezos_protocol_environment_016_PtMumbai.Signature.Public_key_hash.t -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

init ctxt delegate returns a new context from ctxt where the frozen deposits of the implicit contract represented by delegate have been initialized to Tez_repr.zero.

val allocated : Raw_context.t -> Contract_repr.t -> bool Tezos_protocol_environment_016_PtMumbai.Lwt.t

allocated ctxt contract checks whether contract has frozen deposits in ctxt.

val get : Raw_context.t -> Contract_repr.t -> Storage.deposits Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

get ctxt contract retrieves the frozen deposits of contract in ctxt.

val find : Raw_context.t -> Contract_repr.t -> Storage.deposits option Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

find ctxt contract retrieves the frozen deposits of contract in ctxt, if any.

val credit_only_call_from_token : Raw_context.t -> Tezos_protocol_environment_016_PtMumbai.Signature.Public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

credit_only_call_from_token ctxt delegate tez returns a new context from ctxt where the amount of frozen deposits for the implicit contract represented by delegate increases by tez.

val spend_only_call_from_token : Raw_context.t -> Tezos_protocol_environment_016_PtMumbai.Signature.Public_key_hash.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

spend_only_call_from_token ctxt delegate tez returns a new context from ctxt where the amount of frozen deposits for the implicit contract represented by delegate decreases by tez.

val update_initial_amount : Raw_context.t -> Contract_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

update_initial_amount ctxt contract tez returns a new context from ctxt where the initial_amount of the frozen deposits for contract is set to tez.