Module Tezos_raw_protocol_016_PtMumbai.Liquidity_baking_storage

val get_cpmm_address : Raw_context.t -> Contract_hash.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Get the address of the Constant-Product Market Maker receiving the Liquidity Baking subsidy

val on_subsidy_allowed : Raw_context.t -> toggle_vote:Liquidity_baking_repr.liquidity_baking_toggle_vote -> (Raw_context.t -> Contract_hash.t -> (Raw_context.t * 'a list) Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t) -> (Raw_context.t * 'a list * Liquidity_baking_repr.Toggle_EMA.t) Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

on_subsidy_allowed ctxt ~toggle_vote f updates the toggle EMA according to toggle_vote. Then the callback function f is called if the following conditions are met:

the updated EMA is below the threshold,

the CPMM contract exists.

The role of the callback function f is to send the subsidy to the CPMM, see apply_liquidity_baking_subsidy in apply.ml.