Module Tezos_raw_protocol_016_PtMumbai.Sc_rollup_reveal_hash

type t

The type of a reveal hash.

type supported_hashes =

| Blake2B

The hashing schemes supported by the reveal hash.

module Map : Tezos_protocol_environment_016_PtMumbai.Map.S with type key = t

A Map module for storing reveal-hash-indexed values.

val size : scheme:supported_hashes -> int

size ~scheme returns the size of reveal hashes using the scheme specified in input.

val zero : scheme:supported_hashes -> t

zero ~scheme returns the reveal hash corresponding to the zero hash for the scheme specified in input.

val pp : Tezos_protocol_environment_016_PtMumbai.Format.formatter -> t -> unit

Formatting function for reveal-hashes.

val equal : t -> t -> bool

equal hash1 hash2 checks if the two reveal-hashes hash1 and hash2 are equal. This function must preserve the equality of individual supported hashing schemes. If hash1 and hash2 are hashes obtained from the same supported hashing scheme, then the equal function from that hashing scheme is used to determine whether they are equivalent. Otherwise, they are different.

val compare : t -> t -> int

compare hash1 hash2 compares the values of the reveal hashes hash1 and hash2. This function must preserve the ordering of individual supported hashing scheme. If hash1 and hash2 are reveal-hashes obtained from the same hashing scheme, then compare hash1 hash2 should return the same result of the compare function exposed by the hash module corresponding to their hashing scheme.

val of_b58check_opt : string -> t option

val encoding : t Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val hash_string : scheme:supported_hashes -> ?key:string -> string list -> t

val hash_bytes : scheme:supported_hashes -> ?key:bytes -> bytes list -> t

val to_b58check : t -> string

val scheme_of_hash : t -> supported_hashes