Module Sc_rollup_wasm.V2_0_0

This module provides Proof-Generating Virtual Machine (PVM) running WebAssembly (version 2.0.0).

module type S = sig ... end

module type P = sig ... end

module type Make_wasm = module type of Tezos_protocol_environment_016_PtMumbai.Wasm_2_0_0.Make

module Make (Lib_scoru_Wasm : Make_wasm) (Context : P) : S with type context = Context.Tree.t and type state = Context.tree and type proof = Context.proof

Build a WebAssembly PVM using the given proof-supporting context.

module Protocol_implementation : S with type context = Tezos_protocol_environment_016_PtMumbai.Context.t and type state = Tezos_protocol_environment_016_PtMumbai.Context.tree and type proof = Tezos_protocol_environment_016_PtMumbai.Context.Proof.tree Tezos_protocol_environment_016_PtMumbai.Context.Proof.t

This PVM is used for verification in the Protocol. produce_proof always returns None.

val reference_initial_state_hash : Sc_rollup_repr.State_hash.t

This is the state hash of reference that both the prover of the node and the verifier of the protocol Protocol_implementation have to agree on (if they do, it means they are using the same tree structure).

val ticks_per_snapshot : Tezos_protocol_environment_016_PtMumbai.Z.t

Number of ticks between snapshotable states, chosen low enough to maintain refutability.

Warning: This value is used to specialize the dissection predicate of the WASM PVM. Do not change it without a migration plan for already originated smart rollups.

Depends on

speed (tick/s) of node in slow mode (from benchmark, 6000000 tick/s)

the number of ticks in a commitment (Int64.max_int, as per Number_of_ticks.max_value)

see #3590 for more pointers

val outbox_validity_period : int32

val outbox_message_limit : Tezos_protocol_environment_016_PtMumbai.Z.t

val well_known_reveal_hash : Sc_rollup_reveal_hash.t

The hash requested by the WASM PVM if it cannot decode the input provided by the WASM kernel, that is, if the bytes value cannot be decoded with Sc_rollup_reveal_hash.encoding.

val well_known_reveal_preimage : string

The preimage of well_known_reveal_hash.