Module Tezos_raw_protocol_016_PtMumbai.Seed_repr

Tezos Protocol Implementation - Random number generation

This is not expected to be a good cryptographic random number generator. In particular this is supposed to be used in situations where the seed is a globally known information.

The only expected property is: It should be difficult to find a seed such that the generated sequence is a given one.

Random Generation

type t

The state of the random number generator

type seed

A random seed, to derive random sequences from

type sequence

A random sequence, to derive random values from

type vdf_setup = Tezos_protocol_environment_016_PtMumbai.Vdf.discriminant * Tezos_protocol_environment_016_PtMumbai.Vdf.challenge

A VDF discriminant and challenge

type vdf_solution = Tezos_protocol_environment_016_PtMumbai.Vdf.result * Tezos_protocol_environment_016_PtMumbai.Vdf.proof

A VDF result, to derive a seed from

val pp_solution : Tezos_protocol_environment_016_PtMumbai.Format.formatter -> vdf_solution -> unit

val compare_vdf_solution : vdf_solution -> vdf_solution -> int

Compare only the first element of two vdf_solution, that are of Vdf.result.

val generate_vdf_setup : seed_discriminant:seed -> seed_challenge:seed -> vdf_setup

val verify : vdf_setup -> Tezos_protocol_environment_016_PtMumbai.Int64.t -> vdf_solution -> bool option

val vdf_to_seed : seed -> vdf_solution -> seed

val initialize_new : seed -> bytes list -> t

initialize_new state ident returns a new generator

val sequence : t -> int32 -> sequence

sequence state n prepares the n-th sequence of a state

val take : sequence -> bytes * sequence

Generates the next random value in the sequence

val take_int32 : sequence -> int32 -> int32 * sequence

take_int32 s bound generates the next random value as a bounded int32

parameter bound
must be a positive integer

raises Invalid_argument
"Seed_repr.take_int32" if bound <= 0

val take_int64 : sequence -> int64 -> int64 * sequence

take_int64 s bound generates the next random value as a bounded int64

parameter bound
must be a positive integer

raises Invalid_argument
"Seed_repr.take_int64" if bound <= 0

Entropy

type nonce

A nonce for adding entropy to the generator

val update_seed : seed -> nonce -> seed

Add entropy to the seed generator

val make_nonce : bytes -> nonce Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult

Use a byte sequence as a nonce

val hash : nonce -> Nonce_hash.t

Compute the hash of a nonce

val check_hash : nonce -> Nonce_hash.t -> bool

check_hash nonce hash is true if the nonce correspond to the hash

val nonce_hash_key_part : Nonce_hash.t -> string list -> string list

For using nonce hashes as keys in the hierarchical database

val deterministic_seed : seed -> seed

Returns a new seed by hashing the one passed with a constant.

val initial_seeds : ?initial_seed:State_hash.t -> int -> seed list

initial_seeds n generates the first n seeds for which there are no nonces. The first seed is a constant value. The kth seed is the hash of seed (k-1) concatenated with a constant. If an initial_seed is provided, the first seed is created using it as the first one.

Predefined nonce

val initial_nonce_0 : nonce

val initial_nonce_hash_0 : Nonce_hash.t

Serializers

val nonce_encoding : nonce Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val seed_encoding : seed Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val vdf_setup_encoding : vdf_setup Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val vdf_solution_encoding : vdf_solution Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

type seed_status =

| RANDAO_seed

| VDF_seed

val seed_status_encoding : seed_status Tezos_protocol_environment_016_PtMumbai.Data_encoding.t