Module Tezos_raw_protocol_016_PtMumbai.Seed_storage

This modules handles the storage of random nonce seeds.

This module is responsible for maintaining the table Storage.Seed.For_cycle.

type seed_computation_status =

| Nonce_revelation_stage

| Vdf_revelation_stage of {

seed_discriminant : Seed_repr.seed;

seed_challenge : Seed_repr.seed;

}

| Computation_finished

type Tezos_protocol_environment_016_PtMumbai.Error_monad.error +=

| Unknown of {

oldest : Cycle_repr.t;

cycle : Cycle_repr.t;

latest : Cycle_repr.t;

}

| Already_accepted

| Unverified_vdf

| Too_early_revelation

val init : ?initial_seed:State_hash.t -> Raw_context.t -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Generates the first preserved_cycles+2 seeds for which there are no nonces.

val check_vdf : Raw_context.t -> Seed_repr.vdf_solution -> unit Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Verifies if a VDF (result, proof) is valid.

returns
Error Too_early_revelation if the nonce revelation threshold is greater than the current level cycle position.

returns
Error Already_accepted if a VDF seed has already been recorded.

returns
Error Unverified_vdf if the Seed_repr.vdf_solution is not verified.

val update_seed : Raw_context.t -> Seed_repr.vdf_solution -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Updates the seed with a function of the VDF result.

val raw_for_cycle : Raw_context.t -> Cycle_repr.t -> Seed_repr.seed Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Returns the seed associated with the given cycle. Returns a generic storage error when the seed is not available.

val for_cycle : Raw_context.t -> Cycle_repr.t -> Seed_repr.seed Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Returns the seed associated with the given cycle. Returns the Unknown error when the seed is not available.

val compute_randao : Raw_context.t -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Computes RANDAO output for cycle #(current_cycle + preserved + 1)

val cycle_end : Raw_context.t -> Cycle_repr.t -> (Raw_context.t * Nonce_storage.unrevealed list) Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Must be run at the end of the cycle, resets the VDF state and returns unrevealed nonces to know which party has to forfeit its endorsing rewards for that cycle.

val get_seed_computation_status : Raw_context.t -> seed_computation_status Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Return the random seed computation status, that is whether the VDF computation period has started, and if so the information needed, or if it has finished for the current cycle.

val remove_for_cycle : Raw_context.t -> Cycle_repr.t -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Removes the seed associated with the given cycle from the storage. It assumes the seed exists. If it does not it returns a generic storage error.