Module Tezos_raw_protocol_016_PtMumbai.Skip_list_repr

This module provides an implementation of the skip list data structure.

module type MONAD = sig ... end

Basic signature for a monad.

module type S = sig ... end

A skip list represents a sequence of values. There are three main differences between these skip lists and OCaml standard lists:

module Make (_ : sig ... end) : S