Module Storage.Tx_rollup

module State : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Tx_rollup_repr.t and type value = Tx_rollup_state_repr.t and type t := Raw_context.t

State stores the state of a transaction rollup.

module Inbox : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type t := Raw_context.t * Tx_rollup_repr.t and type key = Tx_rollup_level_repr.t and type value = Tx_rollup_inbox_repr.t

The representation of an inbox. See Tx_rollup_inbox_repr.t for a description of the actual content.

module Revealed_withdrawals : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type t := Raw_context.t * Tx_rollup_repr.t and type key = Tx_rollup_level_repr.t and type value = Bitset.t

A carbonated storage of the set of withdrawals revealed of those potentially associated to each message of an inbox. The key is the message number, which is sequentially assigned from 0.

module Commitment : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Tx_rollup_level_repr.t and type value = Tx_rollup_commitment_repr.Submitted_commitment.t and type t := Raw_context.t * Tx_rollup_repr.t

A rollup can have at most one commitment per rollup level. Some metadata are saved in addition to the commitment itself. See Tx_rollup_commitment_repr.Submitted_commitment.t for the exact content.

module Commitment_bond : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Tezos_protocol_environment_016_PtMumbai.Signature.public_key_hash and type value = int and type t := Raw_context.t * Tx_rollup_repr.t

This stores information about which contracts have bonds for each rollup, and how many commitments those bonds stake.