Module Storage_functors.Pair

Parameters

module I1 : INDEX

module I2 : INDEX

Signature

type t = I1.t * I2.t

include Path_encoding.S with type t := t

val to_path : t -> string list -> string list

to_path t postfix returns the context path name for t postfixed with postfix

val of_path : string list -> t option

of_path path parses path as a context path name for t

val path_length : int

Directory levels of the path encoding of t

type 'a ipath

val args : ('a, t, 'a ipath) Storage_description.args