Module Tx_rollup_commitment_repr.Submitted_commitment

type nonrec t = {

commitment : Compact.t;

commitment_hash : Hash.t;

committer : Tezos_protocol_environment_016_PtMumbai.Signature.Public_key_hash.t;

submitted_at : Raw_level_repr.t;

finalized_at : Raw_level_repr.t option;

}

When a commitment is submitted, we store the committer and the block the commitment was submitted_at along with the commitment itself with its hash.

val encoding : t Tezos_protocol_environment_016_PtMumbai.Data_encoding.t