Module Tezos_raw_protocol_016_PtMumbai.Tx_rollup_inbox_repr

module Merkle : sig ... end

See Merkle_List for the documentation of those functions.

type t = {

inbox_length : int;

cumulated_size : int;

merkle_root : Merkle.root;

}

The view of an inbox: stores the cumulated_size in bytes for the inbox, the inbox_length (i.e., the number of messages), and the cumulative hash of the inbox contents. For newly created inboxes, the hash is initialized as an array 32 null byte.

val size : Tezos_protocol_environment_016_PtMumbai.Z.t

size is the number of bytes necessary to store an inbox in the layer-1 storage.

val (=) : t -> t -> bool

val encoding : t Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val empty : t

val pp : Tezos_protocol_environment_016_PtMumbai.Format.formatter -> t -> unit