Module Tx_rollup_l2_apply.Make

Parameters

module Context : Tx_rollup_l2_context_sig.CONTEXT

Signature

type ctxt = Context.t

The operations are versioned (see Tx_rollup_l2_batch), so their interpretations are.

module Batch_V1 : sig ... end

val apply_deposit : ctxt -> Alpha_context.Tx_rollup_message.deposit -> (ctxt * Message_result.deposit_result * Alpha_context.Tx_rollup_withdraw.t option) Context.m

apply_deposit ctxt deposit credits a quantity of tickets to a layer2 address in ctxt.

This function can fail if the deposit.amount is not strictly-positive.

If the deposit causes an error, then a withdrawal returning the funds to the deposit's sender is returned.

val apply_message : ctxt -> parameters -> Alpha_context.Tx_rollup_message.t -> (ctxt * Message_result.t) Context.m

apply_message ctxt parameters message interprets the message in the ctxt.

That is,

Deposit tickets if the message is a deposit.

Decodes the batch and interprets it for the correct batch version.

The function can fail with Invalid_batch_encoding if it's not able to decode the batch.

The function can also return errors from subsequent functions, see apply_deposit and batch interpretations for various versions.

The list of withdrawals in the message result followed the ordering of the contents in the message.