Parameter Make.S

type t

The state of the storage.

The API adopts a functional paradigm, where the set function returns a new state for the storage, and where it should be possible to reuse a previous state.

type 'a m

The monad of the storage backend.

module Syntax : sig ... end

The necessary monadic operators the monad of the storage backend is required to provide.

val get : t -> bytes -> bytes option m

get storage key returns the value stored in storage for key, if it exists. Returns None if it does not.

val set : t -> bytes -> bytes -> t m

set storage key computes a new state for the storage wherein the value associated to key is value.

storage is expected to remain usable and consistent even after the execution of set.

val remove : t -> bytes -> t m

remove storage key removes key from the storage.