Module Tezos_raw_protocol_016_PtMumbai.Tx_rollup_message_repr

Communication from the layer-1 (Tezos) to the layer-2 (a transaction rollup) happens thanks to messages, crafted in the layer-1 to be interpreted in the layer-2.

Messages are constructed and gathered in the layer-1, in inboxes (see Tx_rollup_repr_storage.append_message).

type deposit = {

sender : Tezos_protocol_environment_016_PtMumbai.Signature.Public_key_hash.t;

destination : Tx_rollup_l2_address.Indexable.value;

ticket_hash : Ticket_hash_repr.t;

amount : Tx_rollup_l2_qty.t;

}

Smart contract on the layer-1 can deposit tickets into a transaction rollup, for the benefit of a Tx_rollup_l2_address.t. The sender is an implicit account where the deposit is returned in form of a withdrawal, should the application of the deposit fail.

type t =

| Batch of string

| Deposit of deposit

A message is a piece of data originated from the layer-1 to be interpreted by the layer-2.

Transaction rollups feature two kind of messages:

An array of bytes that supposedly contains a valid sequence of layer-2 operations; their interpretation and validation is deferred to the layer-2..

A deposit order for a L1 ticket.

val size : t -> int

size msg returns the number of bytes that are allocated in an inbox by msg.

val deposit_encoding : deposit Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val encoding : t Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val pp : Tezos_protocol_environment_016_PtMumbai.Format.formatter -> t -> unit