Module Tezos_raw_protocol_016_PtMumbai.Tx_rollup_message_result_repr

type t = {

context_hash : Tezos_protocol_environment_016_PtMumbai.Context_hash.t;

withdraw_list_hash : Tx_rollup_withdraw_list_hash_repr.t;

}

val encoding : t Tezos_protocol_environment_016_PtMumbai.Data_encoding.t

val init : t

val empty_l2_context_hash : Tezos_protocol_environment_016_PtMumbai.Context_hash.t

empty_l2_context_hash is the context hash of the layer-2 context just after its origination.

The empty layer2 context hash is the hash of the underlying Irmin tree. One important note is: an empty tree *must* not be hashed when it's empty. See https://github.com/mirage/irmin/issues/1304.

Our solution is to write data in the tree to have a non-empty one. We write the Tx_rollup_l2_context.key.Ticket_count default value (i.e. 0) and the Tx_rollup_l2_context.key.Address_count as well in the tree. Then we hash the resulting tree to create this constant.