Module Voting_period_storage.Testnet_dictator

val overwrite_current_kind : Raw_context.t -> Tezos_protocol_environment_016_PtMumbai.Chain_id.t -> Voting_period_repr.kind -> Raw_context.t Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult Tezos_protocol_environment_016_PtMumbai.Lwt.t

Overwrites the kind of the current voting period WITHOUT incrementing the index.

Must ONLY be called by the testnet dictator on a testnet.

returns
Error Storage_error if the current voting period is not set or its deserialization fails.