Module Tezos_raw_protocol_016_PtMumbai.Zk_rollup_parameters

A module for representing and extracting typed ZK rollup parameters.

type deposit_parameters = {

ex_ticket : Ticket_scanner.ex_ticket;

zkru_operation : Alpha_context.Zk_rollup.Operation.t;

}

A type representing deposit parameters for ZK rollups. Deposit parameters consist of a ticket of arbitrary content along with a layer-2 ZKRU operation byte representation.

val get_deposit_parameters : (('a Script_typed_ir.ticket, bytes) Script_typed_ir.pair, 'comparable) Script_typed_ir.ty -> ('a Script_typed_ir.ticket, bytes) Script_typed_ir.pair -> deposit_parameters Tezos_protocol_environment_016_PtMumbai.Error_monad.tzresult

get_deposit_parameters ty value returns ex_ticket and a zkru_operation from a michelson typed value. if ty is not of a pair of ticket and bytes then it fails with Zk_rollup_errors.Wrong_deposit_parameters.

This function is intended to be used to enforce the type of the transaction to a zk_rollup%deposit. It must be used both in ticket_diffs_of_operations to account for the ticket deposited and in apply to retrieve the ticket when applying the transaction to a zk_rollup.